Switched to using new constant-time base32 encoder
